# docs/api/openapi.yaml
openapi: 3.2.1
info:
  title: School Report API
  version: 1.0.0
  description: |
    REST API for the School Report MVP for Indonesian elementary schools.
    One deployed application belongs to exactly one school. The authenticated
    session determines school context; clients MUST NOT use schoolId for authorization.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /subjects/{subjectId}/cp:
    post:
      tags: [Curriculum]
      operationId: createCp
      parameters: [{ $ref: '#/components/parameters/SubjectId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateCpRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/CurriculumOutcome' } } } }
  /subjects/{subjectId}/curriculum:
    get:
      tags: [Curriculum]
      operationId: listSubjectCurriculum
      parameters: [{ $ref: '#/components/parameters/SubjectId' }]
      responses:
        '200': { description: Curriculum outcomes with learning objectives (ordered by code), content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/CurriculumOutcomeWithTps' } } } } }
  /cp/{cpId}:
    patch:
      tags: [Curriculum]
      operationId: updateCp
      parameters: [{ $ref: '#/components/parameters/CpId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateCpRequest' } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/CurriculumOutcome' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Curriculum]
      operationId: deleteCp
      parameters: [{ $ref: '#/components/parameters/CpId' }]
      responses:
        '200': { description: Deleted, content: { application/json: { schema: { $ref: '#/components/schemas/OkResult' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /cp/{cpId}/tp:
    post:
      tags: [Curriculum]
      operationId: createTp
      parameters: [{ $ref: '#/components/parameters/CpId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateTpRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/LearningObjective' } } } }
  /tp/{tpId}:
    patch:
      tags: [Curriculum]
      operationId: updateTp
      parameters: [{ $ref: '#/components/parameters/TpId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateTpRequest' } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/LearningObjective' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Curriculum]
      operationId: deleteTp
      parameters: [{ $ref: '#/components/parameters/TpId' }]
      responses:
        '200': { description: Deleted, content: { application/json: { schema: { $ref: '#/components/schemas/OkResult' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /assessments:
    get:
      tags: [Assessments]
      operationId: listAssessments
      parameters:
        - { in: query, name: semesterId, schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: classId, schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: subjectId, schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: categoryId, schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/AssessmentStatus' } }
      responses:
        '200': { description: Assessments, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Assessment' } } } } }
    post:
      tags: [Assessments]
      operationId: createAssessment
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateAssessmentRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Assessment' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /assessments/{assessmentId}:
    get:
      tags: [Assessments]
      operationId: getAssessment
      parameters: [{ $ref: '#/components/parameters/AssessmentId' }]
      responses:
        '200': { description: Assessment, content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentDetail' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /assessments/{assessmentId}/tps:
    put:
      tags: [Assessments]
      operationId: setAssessmentTps
      parameters: [{ $ref: '#/components/parameters/AssessmentId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SetAssessmentTpsRequest' } } }
      responses:
        '200': { description: Assessment, content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentDetail' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /assessments/{assessmentId}/scores:
    get:
      tags: [Assessments]
      operationId: getScores
      parameters: [{ $ref: '#/components/parameters/AssessmentId' }]
      responses:
        '200': { description: Scores, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AssessmentScore' } } } } }
    put:
      tags: [Assessments]
      operationId: replaceScores
      parameters: [{ $ref: '#/components/parameters/AssessmentId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BulkScoresRequest' } } }
      responses:
        '200': { description: Saved transactionally, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AssessmentScore' } } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /assessments/{assessmentId}/scores/import:
    post:
      tags: [Assessments]
      operationId: previewScoreImport
      parameters: [{ $ref: '#/components/parameters/AssessmentId' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { description: Preview, content: { application/json: { schema: { $ref: '#/components/schemas/ImportPreview' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /assessments/{assessmentId}/scores/import/commit:
    post:
      tags: [Assessments]
      operationId: commitScoreImport
      parameters: [{ $ref: '#/components/parameters/AssessmentId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ImportCommitRequest' } } }
      responses:
        '200': { description: Imported, content: { application/json: { schema: { $ref: '#/components/schemas/ImportCommitResult' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: school_report_session
  parameters:
...
    SubjectId: { in: path, name: subjectId, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    CpId: { in: path, name: cpId, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    TpId: { in: path, name: tpId, required: true, schema: { $ref: '#/components/schemas/UUID' } }
...
    AssessmentId: { in: path, name: assessmentId, required: true, schema: { $ref: '#/components/schemas/UUID' } }
...
  responses:
...
    NotFound: { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Conflict: { description: Domain conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    ValidationError: { description: Validation/domain error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  schemas:
    UUID: { type: string, format: uuid }
    Date: { type: string, format: date }
...
    AssessmentStatus: { type: string, enum: [DRAFT, PUBLISHED, CLOSED] }
...
    CurriculumOutcome:
      type: object
      required: [id, subjectId, code, description, isActive]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        subjectId: { $ref: '#/components/schemas/UUID' }
        code: { type: string }
        description: { type: string }
        isActive: { type: boolean }
    CreateCpRequest:
      type: object
      required: [code, description]
      properties:
        code: { type: string }
        description: { type: string }
    LearningObjective:
      type: object
      required: [id, cpId, code, description, isActive]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        cpId: { $ref: '#/components/schemas/UUID' }
        code: { type: string }
        description: { type: string }
        isActive: { type: boolean }
    CreateTpRequest:
      type: object
      required: [code, description]
      properties:
        code: { type: string }
        description: { type: string }
    UpdateCpRequest:
      type: object
      properties:
        code: { type: string }
        description: { type: string }
        isActive: { type: boolean }
    UpdateTpRequest:
      type: object
      properties:
        code: { type: string }
        description: { type: string }
        isActive: { type: boolean }
    CurriculumOutcomeWithTps:
      allOf:
        - $ref: '#/components/schemas/CurriculumOutcome'
        - type: object
          properties:
            tps: { type: array, items: { $ref: '#/components/schemas/LearningObjective' } }
    OkResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Assessment:
      type: object
      required: [id, teacherAssignmentId, semesterId, classId, subjectId, categoryId, title, maxScore, status]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        teacherAssignmentId: { $ref: '#/components/schemas/UUID' }
        semesterId: { $ref: '#/components/schemas/UUID' }
        classId: { $ref: '#/components/schemas/UUID' }
        subjectId: { $ref: '#/components/schemas/UUID' }
        categoryId: { $ref: '#/components/schemas/UUID' }
        title: { type: string }
        description: { type: string, nullable: true }
        assessmentDate: { allOf: [{ $ref: '#/components/schemas/Date' }], nullable: true }
        maxScore: { type: number, exclusiveMinimum: 0 }
        status: { $ref: '#/components/schemas/AssessmentStatus' }
        tps: { type: array, items: { $ref: '#/components/schemas/AssessmentTp' } }
    CreateAssessmentRequest:
      type: object
      required: [teacherAssignmentId, semesterId, classId, subjectId, categoryId, title, maxScore]
      properties:
        teacherAssignmentId: { $ref: '#/components/schemas/UUID' }
        semesterId: { $ref: '#/components/schemas/UUID' }
        classId: { $ref: '#/components/schemas/UUID' }
        subjectId: { $ref: '#/components/schemas/UUID' }
        categoryId: { $ref: '#/components/schemas/UUID' }
        title: { type: string, minLength: 1 }
        description: { type: string }
        assessmentDate: { $ref: '#/components/schemas/Date' }
        maxScore: { type: number, exclusiveMinimum: 0 }
        tpIds: { type: array, items: { $ref: '#/components/schemas/UUID' } }
    AssessmentTp:
      type: object
      required: [id, code, description, cpId]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        code: { type: string }
        description: { type: string }
        cpId: { $ref: '#/components/schemas/UUID' }
    SetAssessmentTpsRequest:
      type: object
      required: [tpIds]
      properties:
        tpIds: { type: array, items: { $ref: '#/components/schemas/UUID' } }
    AssessmentScore:
      type: object
      required: [id, assessmentId, studentId, score, normalizedScore]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        assessmentId: { $ref: '#/components/schemas/UUID' }
        studentId: { $ref: '#/components/schemas/UUID' }
        score: { type: number, minimum: 0 }
        normalizedScore: { type: number, minimum: 0, maximum: 100 }
        note: { type: string, nullable: true }
    AssessmentDetail:
      allOf:
        - $ref: '#/components/schemas/Assessment'
        - type: object
          properties:
            scores: { type: array, items: { $ref: '#/components/schemas/AssessmentScore' } }
    BulkScoresRequest:
      type: object
      required: [scores]
      properties:
        scores:
          type: array
          items:
            type: object
            required: [studentId, score]
            properties:
              studentId: { $ref: '#/components/schemas/UUID' }
              score: { type: number, minimum: 0 }
              note: { type: string }
    ImportPreview:
      type: object
      required: [importId, valid, totalRows, validRows, invalidRows, errors]
      properties:
        importId: { $ref: '#/components/schemas/UUID' }
        valid: { type: boolean }
        totalRows: { type: integer, minimum: 0 }
        validRows: { type: integer, minimum: 0 }
        invalidRows: { type: integer, minimum: 0 }
        errors: { type: array, items: { $ref: '#/components/schemas/ImportRowError' } }
    ImportRowError:
      type: object
      required: [row, code, message]
      properties:
        row: { type: integer, minimum: 1 }
        code: { type: string }
        message: { type: string }
    ImportCommitRequest:
      type: object
      required: [importId]
      properties:
        importId: { $ref: '#/components/schemas/UUID' }
    ImportCommitResult:
      type: object
      required: [importId, importedRows]
      properties:
        importId: { $ref: '#/components/schemas/UUID' }
        importedRows: { type: integer, minimum: 0 }
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - AUTH_INVALID_CREDENTIALS
            - AUTH_SESSION_EXPIRED
            - AUTH_FORBIDDEN
            - VALIDATION_ERROR
            - RESOURCE_NOT_FOUND
            - RESOURCE_CONFLICT
            - STUDENT_ALREADY_ENROLLED
            - INVALID_ACADEMIC_YEAR
            - INVALID_SEMESTER
            - INVALID_TEACHER_ASSIGNMENT
            - INVALID_WEIGHT_TOTAL
            - GRADING_SCHEME_PUBLISHED
            - ASSESSMENT_SCORE_EXCEEDED_MAX
            - ASSESSMENT_INCOMPLETE
            - REPORT_NOT_READY
            - REPORT_ALREADY_PUBLISHED
            - REPORT_NOT_LOCKED
            - REPORT_REVISION_REQUIRED
            - IMPORT_INVALID_FORMAT
            - IMPORT_VALIDATION_FAILED
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { type: string }